:zap: :: change snap api response dto
